renames asset data provider
