+++ apps/gp2-server/src/data-providers/asset.data-provider.ts
