+++ apps/gp2-server/src/handlers/webhooks/fetch-by-code-handler.ts
@@ -9,7 +9,7 @@ import {
 } from '@asap-hub/squidex';
 import { appName, auth0SharedSecret, baseUrl } from '../../config';
 import Users, { UserController } from '../../controllers/user.controller';
-import { AssetSquidexDataProvider } from '../../data-providers/assets.data-provider';
+import { AssetSquidexDataProvider } from '../../data-providers/asset.data-provider';
 import { UserSquidexDataProvider } from '../../data-providers/user.data-provider';
 import { getAuthToken } from '../../utils/auth';
 import { sentryWrapper } from '../../utils/sentry-wrapper';
