+++ apps/gp2-server/test/mocks/asset-data-provider.mock.ts
@@ -1,4 +1,4 @@
-import { AssetDataProvider } from '../../src/data-providers/assets.data-provider';
+import { AssetDataProvider } from '../../src/data-providers/asset.data-provider';
 
 export const assetDataProviderMock: jest.Mocked<AssetDataProvider> = {
   create: jest.fn(),
